Clean up FrameTimeDisplay; split out classes.
